% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_models.R
\name{assessModelMetrics}
\alias{assessModelMetrics}
\title{Invoke calculation of performance metrics per model}
\usage{
assessModelMetrics(fits, x, groups)
}
\description{
Invoke calculation of performance metrics per model
}
