% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_models.R
\name{augmentModels}
\alias{augmentModels}
\title{Augment model to obtain data frame with measurements, predictions and residuals.}
\usage{
augmentModels(fits, groups)
}
\description{
Augment model to obtain data frame with measurements, predictions and residuals.
}
