% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{augmentSingleModel}
\alias{augmentSingleModel}
\title{Compute predictions and residuals for a single model}
\usage{
augmentSingleModel(nls_obj)
}
\description{
Compute predictions and residuals for a single model
}
