% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{getParams}
\alias{getParams}
\title{Control parameters for model fitting}
\usage{
getParams(start = c(Pl = 0, a = 550, b = 10), seed = 123,
  maxAttempts = 100)
}
\arguments{
\item{start}{Numeric vector of start parameters for the melting curve equation}

\item{seed}{Random seed to control resampling in case of unsuccessful model fits}

\item{maxAttempts}{Number of resampling steps in case of unsuccessful model fits}
}
\description{
Control parameters for model fitting
}
\examples{
data(stauro_TPP_data_tidy)
df <- dplyr::filter(stauro_TPP_data_tidy, grepl("MAPK|ATP|CDK|GTP|CRK", uniqueID))
testResults <- runNPARC(x = df$temperature,
                     y = df$relAbundance,
                     id = df$uniqueID,
                     groupsAlt = df$compoundConcentration,
                     df_type = "empirical",
                     control = getParams(maxAttempts = 50))

}
