% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPARC.R
\docType{package}
\name{NPARC}
\alias{NPARC}
\alias{NPARC-package}
\title{\code{NPARC} package}
\description{
Non-parametric analysis of response curves
}
\details{
See the preprint on
\href{https://www.biorxiv.org/content/10.1101/373845v2}{Childs, Bach, Franken et al. (2019): Non-parametric analysis of thermal proteome profiles reveals novel drug-binding proteins}
}
