% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{NPARCtest}
\alias{NPARCtest}
\title{Perform F-test}
\usage{
NPARCtest(modelMetrics, df_type = c("empirical", "theoretical"))
}
\arguments{
\item{modelMetrics}{data.frame with results of the model fit in long format.}

\item{df_type}{character value indicating the method for degrees of freedom computation for the F-test. Theoretical yields the text-book solution. Empirical yields estimates derived from the distribution moments of the RSS.}
}
\description{
Perform F-test
}
