% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{repeatSingleFit}
\alias{repeatSingleFit}
\title{Wrapper to repeat the fit until model has converged}
\usage{
repeatSingleFit(x, y, start, seed = NULL, maxAttempts = 100)
}
\arguments{
\item{seed}{Random seed to control resampling in case of unsuccessful model fits}

\item{maxAttemps}{number of resampling steps in case of unsuccessful model fits}
}
\description{
Wrapper to repeat the fit until model has converged
}
