% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stauro_TPP_data_tidy.R
\docType{data}
\name{stauro_TPP_data_tidy}
\alias{stauro_TPP_data_tidy}
\title{TPP dataset of staurosporine treated cells.}
\format{An object of class \code{"data.frame"}}
\usage{
data(stauro_TPP_data_tidy)
}
\description{
Data from a thermal proteome profiling (TPP) experiment investigating the
ATP-competitive pan-kinase inhibitor staurosporine on K562 cells.
The data has been downloaded the data from the supplement of the respective publication and converted into tidy format.
}
\references{
Savitski et al. (2014): Tracking cancer drugs in living cells by thermal profiling of the proteome. Science 346, 1255784.
}
\keyword{datasets}
