% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nparc}
\alias{nparc}
\title{Main function}
\usage{
nparc(x, y, group, id = NULL, params)
}
\description{
Main function
}
\examples{
data(staurosporineTPP)
nparc(x = staurosporineTPP$temperature,
      y = staurosporineTPP$relAbundance,
      group = staurosporineTPP$compoundConcentration)

}
