% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nparc}
\alias{nparc}
\title{Main function}
\usage{
nparc(x, y, id, groupsNull = NULL, groupsAlt,
  BPPARAM = BiocParallel::SerialParam(progressbar = TRUE), seed = 123,
  return_models = FALSE, maxAttempts = 100, alwaysPermute = TRUE,
  df_type = c("theoretical", "estimate"))
}
\description{
Main function
}
\examples{
data(staurosporineTPP)
nparc(x = staurosporineTPP$temperature,
      y = staurosporineTPP$relAbundance,
      id = staurosporineTPP$uniqueID,
      groupsAlt = staurosporineTPP$compoundConcentration)

}
