% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{nparcFit}
\alias{nparcFit}
\title{Fit null and alternative models}
\usage{
nparcFit(x, y, id, groupsNull = NULL, groupsAlt,
  BPPARAM = BiocParallel::SerialParam(progressbar = TRUE), seed = 123,
  return_models = FALSE, verbose = FALSE, maxAttempts = 100,
  alwaysPermute = TRUE)
}
\description{
Fit null and alternative models
}
